Cambio del tipo de dato CapacidadWatts a Decimal para concordancia con ConsumoWatts

+++ backend/ArmatuXPC.Backend/Models/Componente.cs
@@ -22,7 +22,7 @@ namespace ArmatuXPC.Backend.Models
 
         // Solo para fuente de poder
         [Column(TypeName = "decimal(10,2)")]
-        public int? CapacidadWatts { get; set; } // Potencia total en watts que puede suministrar la fuente de poder, para validar que sea suficiente para el armado
+        public decimal? CapacidadWatts { get; set; } // Potencia total en watts que puede suministrar la fuente de poder, para validar que sea suficiente para el armado
         
         // Imagen del componente (URL o base64), para mostrar en el frontend
         public string? ImagenUrl { get; set; }
